requirements.txt
typer==0.9.0
PyYAML==6.0.1
parse==1.20.0
mergedeep==1.3.4
pymysql==1.1.0'
psycopg2==2.9.9
psycopg2-binary==2.9.9
pymssql==2.2.10